Return entire flags hash
As a library, the ActiveBugzilla gem should not assume that all callers only
want a subset of flag data.

Now, the ActiveBugzilla gem will return all of the available flag data as an
array of hashes.

Previously a bug's flags would appear as:

  flags = {"devel_ack" => "?", "qa_ack" => "?"}

Now, a bug's flags will appear as:

  flags = [{"is_active"=>1, "status"=>"?", "type_id"=>10, "name"=>"devel_ack",
            "id"=>2207657, "setter"=>"setter@email.com", "bug_id"=>nil,
            "created_on"=>2015-04-10 13:16:47 UTC,
            "updated_on"=>2015-04-10 13:16:47 UTC},
           {"is_active"=>1, "status"=>"?", "type_id"=>9, "name"=>"qa_ack",
            "id"=>2207662, "setter"=>"setter@email.com", "bug_id"=>nil,
            "created_on"=>2015-04-10 13:17:08 UTC,
            "updated_on" =>2015-04-10 13:17:08 UTC}]

This will allow different clients to choose how much flag data is important.